refactor: add new folders for better segregations
