+++ src/components/pages/Main/Config/SliderSection.tsx
