+++ src/components/pages/Main/Duration/Duration.tsx
