+++ src/components/pages/Main/Salary/Salary.tsx
